Ejercicio para subir

+++ EjercicioUnidad1/src/main/java/Pagos/IngresodePagos.java
@@ -1,7 +1,7 @@
 package Pagos;
 
 public interface IngresodePagos {
-    public void Autorizar(int MontoAAutorizar);
-    public void Capturar(int MontoACapturar);
+    public void Autorizar(double MontoAAutorizar);
+    public void Capturar(double MontoACapturar);
 
 }
