+++ EjercicioUnidad1/src/main/java/Pagos/ManejadordePagos.java
@@ -1,7 +1,7 @@
 package Pagos;
 
 public class ManejadordePagos {
-    public void ProcesadorPagos(int MontoAPagar, TiposPago tipodePago) {
+    public void ProcesadorPagos(double MontoAPagar, TiposPago tipodePago) {
         if (tipodePago==TiposPago.MERCADOPAGO) {
             System.out.println("realizando un pago de" + MontoAPagar + "pesos con MercadoPago");
         }
